requirements.txt
Flask==3.0.0
gunicorn==21.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
python-docx==1.1.0